Move s3_to_postgres from s3 callables to sql callables; make s3.list_s3_keys() public.
